mad!: compute MAD without copying the array, only permuting it.
mad! *only* permutes the values of the input array, it doesn't do
anything else -- the obvious mutation version of this would replace
each value with its absolute deviation, but that somehow seems like
a much worse mutation than just reordering the values. This is less
performant than the version that copies since it uses anonymous
functions for selection.