remove warrays
Reason: I found this solution inadequate, which is only suitable for cases where the weights are associated with only a single vector. There are cases where a set of weights are associated with several arrays or slices of an array. In such cases, the weighted vector is useless. Therefore, it still seems to me that an additional argument for weights is the sane way forward.