Handle missing values and empty inputs in SummaryStats (#465)
* Handle missing values and empty inputs in SummaryStats

Currently, `SummaryStats` will error when computing quantiles if the
input is empty or contains any missing values. Instead, we can simply
inform the user that the input is empty or that there were missing
values that caused the statistics to be missing.

* Add a Project.toml file

They no longer mess up METADATA checks and they're immensely useful for
local development.