Add information field to ConvergenceException
In order to add informative hints to ConvergenceExceptions, I have added a
`info` field to `ConvergenceException` structs. If the `info` field is not
empty, it will also be printed with the error message for
`ConvergenceExceptions`. This is useful to inform the user of why he or she may
be experiencing problems with convergence, and to suggest ways to avoid these
errors. For an example of this, see JuliaStats/GLM.jl#303.