Improve printing CoefTable (#481)
Rely on printing code from Base to align numbers on the decimal separator.
Unfortunately, we cannot treat column names as a normal row since names would
be left-aligned with the decimal separator, which wastes space and is less clean.
Instead, adjust the alignment manually so that columns are wide enough to contain
names.

Only convert values to PValue before printing, so that users can still access the contents
of columns as normal numbers. This is convenient in particular for testing in packages.

Print test statistics with 2 decimals.

Add horizontal lines.