Add a no-op PValue(::PValue) constructor
`show(::CoefTable)` calls `PValue` on all values in the p-value column
(if present). If things are constructed manually and you happen to
already have your p-values as `PValue`s, printing the table will fail.
It seems sort of silly to have this kind of definition, but it's nice to
not get errors when trying to look at your `CoefTable` in the REPL.