Make CoefTable implement the Tables.jl interface
This allows retrieving the contents of a `CoefTable` object in a convenient form,
notably a `DataFrame`. To avoid introducing a dependency on Tables.jl, `CoefTable`
has to iterate `NamedTuple`s, so that it implements the row-table interface
implicitly. This is inefficient since `CoefTable` uses a column-based storage,
but given the typical size of such tables it should not matter.