Reexport functions from Statistics
It does not make any sense to avoid exporting functions from Statistics
to which we add methods. These are documented in the manual but currently
not available without typing `using Statistics`.
The only method from Statistics which we do not overload is `middle`, which
this commit does not reexport for now.