Deprecate `varm` and `stdm`
`varm` and `stdm` are redundant with `var(..., mean=m)` and `std(..., mean=m)`.
Moreover, their status is not clear: they are technically exported by Statistics
(but not by StatsBase itself), but their docstrings are not included in the manual.
They were only tested indirectly, so add tests to ensure deprecations work.

The status of `varm!` is even weirder as it's imported from Statistics, but
Statistics does not export it, and it has no docstrings. Keep a deprecation
just in case.