Fix pairwise for type-unstable corner case function (#772)
`promote_type` is not a completely correct way of computing an upper bound
for the return eltype. Use the same strategy as `map` and `broadcast` in
Base instead, but with `promote_eltype` rather than `promote_typejoin`.
We can still use `typejoin_union_tuple` since promotion does not happen
inside tuple types.

On Julia versions before 1.6 we would have to copy the full definition of
`typejoin_union_tuple`, which is quite complex, so instead fall back to
inferring eltype `Any`.