Fix handling of -0.0 in histograms (#768)
`searchsortedfirst` and `searchsortedlast` use `isless` for comparisons
and therefore consider `-0.0` to be different from `0.0`. This means
that these two values do not end up in the same bin when an edge is 0.
This does not make much sense statistically, but even worse is that
when an extreme edge is 0, `-0.0` is not counted at all.

Fix this by replacing `-0.0` with `0.0` before the search.