Check that inputs and outputs do not share memory in `sample!`
Otherwise the result is incorrect. Unfortunately, there is no completely
reliable way to check whether two arrays alias, but this is the best effort
we can make.

For custom array types like `AbstractWeights`, `Base.mightalias` falls back to
using `Base.dataids`, which is less smart than the `SubArray` method,
generating false positives when passing two views with the same parent
but disjoint indices. In practice this is unlikely to matter, as
using the same argument for values and weights is weird.