Fix tests on Julia nightly
The tests for specific exception types in `sem` are relying on the
exceptions thrown from deeper in the call stack, outside of this
package's control. We do still expect these cases to throw, but since we
can't actually control the type of the exception, we can instead test
that any exception is thrown.

The test of the number of allocations from calling `mad` on a generator
fails because there are nearly 10x more allocations than the upper bound
used in the test. We can mark this test broken with Julia v1.10, and the
test will fail if the allocations for this case are improved, be it
upstream or in the implementation of `mad` (if applicable).