Fix `sample` with wide integer types
As observed in issue 872, an unordered sample without replacement with a
type 64 bits or wider (notably excluding `Int64`, at least on a 64-bit
system) hits a `MethodError` in `samplepair`, which had been restricted
to `Int` inputs. Relaxing this restriction and adjusting the
`samplepair` implementation accordingly allows these cases to succeed.