Unify argument checking for sampling and check more cases
Currently sampling functions need to perform the same set of checks on
the inputs and those checks are copied and pasted for each method. We
can instead define a simple input validation function that can be used
by all sampling functions so that any additional corner cases that need
to be caught can be fixed in one place and propagated elsewhere.

Relatedly, this adds checks for agreement between the length of the
source array to be sampled and the array of weights (issue 871) as well
as that the destination array is not larger than the source when
sampling without replacement (issue 877).