Allow non-`Real` eltype with `quantile` (#981)
This is consistent with `Statistics.quantile` and avoids breakage due to #977.
It's particularly useful for `Union{T, Missing}`, e.g. a view of nonmissing entries
in a vector. This also allows supporting some types such as `Date`, though currently
this only works for some values (would need to implement `type=1`).